+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* [ADD] Add `PrefixedIds.delimiter` to be able to change the default delimiter - @rbague
 * [FIX] Custom alphabet was not being used to generate the prefixed_id - @rbague
 
 ### 1.2.0